# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/payment-intents/{id}": {
      "get": {
        "operationId": "PaymentIntentController.get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "pattern": "^pi_[0-7][0-9A-HJKMNP-TV-Z]{25}$",
              "type": "string"
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "description": "Optional caller correlation ID. Invalid values are replaced. The canonical value is returned in X-Request-Id.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Merchant-owned Payment Intent.",
            "headers": {
              "X-Request-Id": {
                "description": "Canonical correlation ID for this HTTP attempt.",
                "schema": {
                  "example": "req_example-safe-value",
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "amountMinor": {
                      "example": 125000,
                      "maximum": 9007199254740991,
                      "minimum": 1,
                      "type": "integer"
                    },
                    "captureMethod": {
                      "enum": [
                        "manual"
                      ],
                      "example": "manual",
                      "type": "string"
                    },
                    "capturedAmountMinor": {
                      "example": 0,
                      "maximum": 9007199254740991,
                      "minimum": 0,
                      "type": "integer"
                    },
                    "createdAt": {
                      "example": "2026-08-01T10:20:12.345Z",
                      "format": "date-time",
                      "type": "string"
                    },
                    "currency": {
                      "enum": [
                        "ETB",
                        "USD"
                      ],
                      "example": "ETB",
                      "type": "string"
                    },
                    "externalRef": {
                      "example": "order_1001",
                      "maxLength": 255,
                      "minLength": 1,
                      "type": "string"
                    },
                    "id": {
                      "example": "pi_01ARZ3NDEKTSV4RRFFQ69G5FAV",
                      "pattern": "^pi_[0-7][0-9A-HJKMNP-TV-Z]{25}$",
                      "type": "string"
                    },
                    "paymentStatus": {
                      "enum": [
                        "created"
                      ],
                      "example": "created",
                      "type": "string"
                    },
                    "refundedAmountMinor": {
                      "example": 0,
                      "maximum": 9007199254740991,
                      "minimum": 0,
                      "type": "integer"
                    },
                    "settlementStatus": {
                      "enum": [
                        "NOT_ELIGIBLE"
                      ],
                      "example": "NOT_ELIGIBLE",
                      "type": "string"
                    },
                    "updatedAt": {
                      "example": "2026-08-01T10:20:12.345Z",
                      "format": "date-time",
                      "type": "string"
                    },
                    "version": {
                      "example": 0,
                      "minimum": 0,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "amountMinor",
                    "captureMethod",
                    "capturedAmountMinor",
                    "createdAt",
                    "currency",
                    "externalRef",
                    "id",
                    "paymentStatus",
                    "refundedAmountMinor",
                    "settlementStatus",
                    "updatedAt",
                    "version"
                  ],
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "example": "invalid_request",
                      "type": "string"
                    },
                    "detail": {
                      "example": "The request is invalid.",
                      "type": "string"
                    },
                    "requestId": {
                      "example": "req_example-safe-value",
                      "type": "string"
                    },
                    "status": {
                      "example": 400,
                      "type": "integer"
                    },
                    "title": {
                      "example": "Invalid request",
                      "type": "string"
                    },
                    "type": {
                      "example": "https://docs.settleflow.dev/problems/invalid_request",
                      "format": "uri",
                      "type": "string"
                    },
                    "violations": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "reason": {
                            "enum": [
                              "invalid"
                            ],
                            "type": "string"
                          }
                        },
                        "required": [
                          "field",
                          "reason"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "code",
                    "detail",
                    "requestId",
                    "status",
                    "title",
                    "type"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Payment identifier is invalid.",
            "headers": {
              "X-Request-Id": {
                "description": "Canonical correlation ID for this HTTP attempt.",
                "schema": {
                  "example": "req_example-safe-value",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "example": "invalid_request",
                      "type": "string"
                    },
                    "detail": {
                      "example": "The request is invalid.",
                      "type": "string"
                    },
                    "requestId": {
                      "example": "req_example-safe-value",
                      "type": "string"
                    },
                    "status": {
                      "example": 400,
                      "type": "integer"
                    },
                    "title": {
                      "example": "Invalid request",
                      "type": "string"
                    },
                    "type": {
                      "example": "https://docs.settleflow.dev/problems/invalid_request",
                      "format": "uri",
                      "type": "string"
                    },
                    "violations": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "reason": {
                            "enum": [
                              "invalid"
                            ],
                            "type": "string"
                          }
                        },
                        "required": [
                          "field",
                          "reason"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "code",
                    "detail",
                    "requestId",
                    "status",
                    "title",
                    "type"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Merchant API key missing or invalid.",
            "headers": {
              "X-Request-Id": {
                "description": "Canonical correlation ID for this HTTP attempt.",
                "schema": {
                  "example": "req_example-safe-value",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "example": "invalid_request",
                      "type": "string"
                    },
                    "detail": {
                      "example": "The request is invalid.",
                      "type": "string"
                    },
                    "requestId": {
                      "example": "req_example-safe-value",
                      "type": "string"
                    },
                    "status": {
                      "example": 400,
                      "type": "integer"
                    },
                    "title": {
                      "example": "Invalid request",
                      "type": "string"
                    },
                    "type": {
                      "example": "https://docs.settleflow.dev/problems/invalid_request",
                      "format": "uri",
                      "type": "string"
                    },
                    "violations": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "reason": {
                            "enum": [
                              "invalid"
                            ],
                            "type": "string"
                          }
                        },
                        "required": [
                          "field",
                          "reason"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "code",
                    "detail",
                    "requestId",
                    "status",
                    "title",
                    "type"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "API key lacks payments:read.",
            "headers": {
              "X-Request-Id": {
                "description": "Canonical correlation ID for this HTTP attempt.",
                "schema": {
                  "example": "req_example-safe-value",
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "example": "invalid_request",
                      "type": "string"
                    },
                    "detail": {
                      "example": "The request is invalid.",
                      "type": "string"
                    },
                    "requestId": {
                      "example": "req_example-safe-value",
                      "type": "string"
                    },
                    "status": {
                      "example": 400,
                      "type": "integer"
                    },
                    "title": {
                      "example": "Invalid request",
                      "type": "string"
                    },
                    "type": {
                      "example": "https://docs.settleflow.dev/problems/invalid_request",
                      "format": "uri",
                      "type": "string"
                    },
                    "violations": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "reason": {
                            "enum": [
                              "invalid"
                            ],
                            "type": "string"
                          }
                        },
                        "required": [
                          "field",
                          "reason"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "code",
                    "detail",
                    "requestId",
                    "status",
                    "title",
                    "type"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Payment Intent missing or owned by another merchant.",
            "headers": {
              "X-Request-Id": {
                "description": "Canonical correlation ID for this HTTP attempt.",
                "schema": {
                  "example": "req_example-safe-value",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "example": "invalid_request",
                      "type": "string"
                    },
                    "detail": {
                      "example": "The request is invalid.",
                      "type": "string"
                    },
                    "requestId": {
                      "example": "req_example-safe-value",
                      "type": "string"
                    },
                    "status": {
                      "example": 400,
                      "type": "integer"
                    },
                    "title": {
                      "example": "Invalid request",
                      "type": "string"
                    },
                    "type": {
                      "example": "https://docs.settleflow.dev/problems/invalid_request",
                      "format": "uri",
                      "type": "string"
                    },
                    "violations": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "reason": {
                            "enum": [
                              "invalid"
                            ],
                            "type": "string"
                          }
                        },
                        "required": [
                          "field",
                          "reason"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "code",
                    "detail",
                    "requestId",
                    "status",
                    "title",
                    "type"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Unexpected internal failure.",
            "headers": {
              "X-Request-Id": {
                "description": "Canonical correlation ID for this HTTP attempt.",
                "schema": {
                  "example": "req_example-safe-value",
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "example": "invalid_request",
                      "type": "string"
                    },
                    "detail": {
                      "example": "The request is invalid.",
                      "type": "string"
                    },
                    "requestId": {
                      "example": "req_example-safe-value",
                      "type": "string"
                    },
                    "status": {
                      "example": 400,
                      "type": "integer"
                    },
                    "title": {
                      "example": "Invalid request",
                      "type": "string"
                    },
                    "type": {
                      "example": "https://docs.settleflow.dev/problems/invalid_request",
                      "format": "uri",
                      "type": "string"
                    },
                    "violations": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "reason": {
                            "enum": [
                              "invalid"
                            ],
                            "type": "string"
                          }
                        },
                        "required": [
                          "field",
                          "reason"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "code",
                    "detail",
                    "requestId",
                    "status",
                    "title",
                    "type"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Required database unavailable.",
            "headers": {
              "X-Request-Id": {
                "description": "Canonical correlation ID for this HTTP attempt.",
                "schema": {
                  "example": "req_example-safe-value",
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "merchantApiKey": []
          }
        ],
        "summary": "Retrieve one merchant-owned Payment Intent",
        "tags": [
          "payment-intents"
        ],
        "x-required-scopes": [
          "payments:read"
        ]
      }
    }
  },
  "info": {
    "title": "SettleFlow API",
    "description": "Finance-grade payment-platform simulation. Merchant API routes use scoped API keys; no real funds or regulated payment data are supported.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "merchantApiKey": {
        "scheme": "bearer",
        "bearerFormat": "SettleFlow merchant API key",
        "description": "Use Authorization: Bearer <merchant_api_key>. Never place a real key in documentation.",
        "type": "http"
      }
    },
...
  }
}